requirements.txt
boto3==1.35.64
botocore==1.35.99
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
